Phase 1 of restoring Linux build

+++ dependencies/agg/svg/agg_svg_parser.cpp
@@ -282,7 +282,7 @@ namespace svg
                 sprintf(msg,
                     "%s at line %d\n",
                     XML_ErrorString(XML_GetErrorCode(p)),
-                    XML_GetCurrentLineNumber(p));
+                    (int) XML_GetCurrentLineNumber(p));
                 throw exception(msg);
             }
         }
@@ -319,7 +319,7 @@ namespace svg
             sprintf(msg,
                 "%s at line %d\n",
                 XML_ErrorString(XML_GetErrorCode(p)),
-                XML_GetCurrentLineNumber(p));
+                (int) XML_GetCurrentLineNumber(p));
             throw exception(msg);
         }
         XML_ParserFree(p);
